fix: backend persistence, schema fixes, remove root dashboard
- Fix persistence.py: rewrite with correct save_meeting method and
  proper save_speaker_embedding signature (was orphaned after save_insights)
- Add save_meeting() to create a Meeting row on WebSocket start so
  meeting_events FK constraint is satisfied
- Wire save_meeting() call in main.py WebSocket start handler
- Fix asyncpg SSL params: strip sslmode/channel_binding from URL
- Fix BigInteger for speech_start_ms/speech_end_ms/duration_ms columns
- Fix schemas/models/routers to match actual DB columns
  (review_status, quote, confidence; remove status/priority)
- Fix seed_data FK ordering with db.flush() after user and meetings
- Fix diarization service: stub speechbrain.integrations.k2_fsa to
  prevent 503 on pipeline load
- Remove root-level Next.js dashboard (src/, public/, config files)

+++ backend/app/models.py
@@ -1,6 +1,6 @@
 import datetime
 from datetime import timezone
-from sqlalchemy import Column, String, Integer, Float, Boolean, DateTime, ForeignKey, Text
+from sqlalchemy import Column, String, Integer, BigInteger, Float, Boolean, DateTime, ForeignKey, Text
 from sqlalchemy.orm import relationship
 from .database import Base
 
@@ -54,9 +54,9 @@ class MeetingEvent(Base):
 
     speaker_label = Column(String, nullable=True)
     text = Column(Text, nullable=True)
-    speech_start_ms = Column(Integer, nullable=True)
-    speech_end_ms = Column(Integer, nullable=True)
-    duration_ms = Column(Integer, nullable=True)
+    speech_start_ms = Column(BigInteger, nullable=True)
+    speech_end_ms = Column(BigInteger, nullable=True)
+    duration_ms = Column(BigInteger, nullable=True)
     transcription_confidence = Column(Float, nullable=True)
     speaker_confidence = Column(Float, nullable=True)
     language = Column(String, nullable=True)
